BREAKING feat(db): change recommended value of DESECSLAVE_ID
Previously, slave server-id started with 2, followed by the number
configured in .env. Now, it is just the number from .env (which is
recommended to be the integer representation of the machine IPv4
address).

IMPORTANT: Make sure to update the setting in .env so that the
server-id does not collide with other server instances!