feat(replicator): maintain zone list using serials/ endpoint
Previously, the contents of the catalog.internal zone were used to
determine which domains have been added and deleted. We can save
the corresponding AXFRs and logic by using the information from the
serials/ API endpoint for this purpose. We have been querying it
anyway to learn about zone modifications.

This is a prerequisite for switching to the lmdb backend. The reason
is that the previous approach required retrieving the local catalog
serial from the local ns API zone endpoint, which currently is buggy
in the lmdb backend, see https://github.com/PowerDNS/pdns/pull/8707.
The new approach does not rely on this anymore.